+++ src/main/java/lonestarrr/arconia/common/block/Hat.java
@@ -74,7 +74,7 @@ public class Hat extends BaseEntityBlock {
         ItemStack itemUsed = player.getInventory().getSelected();
 
         BlockEntity be = level.getBlockEntity(pos);
-        if (be == null || !(be instanceof HatBlockEntity)) {
+        if (!(be instanceof HatBlockEntity)) {
             return InteractionResult.PASS;
         }
 
@@ -97,10 +97,6 @@ public class Hat extends BaseEntityBlock {
             } else if (!hbe.getResourceGenerated().isEmpty()) {
                 player.sendSystemMessage(Component.translatable("arconia.block.hat.resource_already_set"));
                 return InteractionResult.FAIL;
-            } else if (potbe.getTier() == null || root.getTier().getTier() > potbe.getTier().getTier()) {
-                player.sendSystemMessage(Component.translatable("arconia.block.hat.resource_tier_too_high"));
-                return InteractionResult.FAIL;
-
             } else {
                 // Ok ok ok, let's attempt set the resource already.
                 ItemStack resource = ColoredRoot.getResourceItem(itemUsed);